I finished

+++ dmitriy/src/main/hw/hw2/TaskFour.java
@@ -12,7 +12,7 @@ public class TaskFour {
     public static void main(String[] args) {
         int a = 15;
         int b = 4;
-        float c = a / b;
+        float c = (float)a / b;
         double d = a * 1e-3 + c;
         System.out.println(d);
     }
